+++ core/src/main/java/xdi2/core/features/nodetypes/XdiAbstractClass.java
@@ -124,7 +124,7 @@ public abstract class XdiAbstractClass<U extends XdiInstanceUnordered, O extends
 	 * Returns the number of XDI instances in this XDI class.
 	 */
 	@Override
-	public int getXdiInstancesUnorderedCount() {
+	public long getXdiInstancesUnorderedCount() {
 
 		return new IteratorCounter(this.getXdiInstancesUnordered()).count();
 	}
@@ -134,11 +134,11 @@ public abstract class XdiAbstractClass<U extends XdiInstanceUnordered, O extends
 	 * @return The XDI element.
 	 */
 	@Override
-	public O setXdiInstanceOrdered(int index) {
+	public O setXdiInstanceOrdered(long index) {
 
 		if (index < 0) index = this.getXdiInstancesOrderedCount();
 
-		XDI3SubSegment arcXri = XdiAbstractInstanceOrdered.createArcXri(Integer.toString(index));
+		XDI3SubSegment arcXri = XdiAbstractInstanceOrdered.createArcXri(Long.toString(index));
 
 		ContextNode contextNode = this.getContextNode().getContextNode(arcXri);
 		if (contextNode == null) contextNode = this.getContextNode().createContextNode(arcXri);
@@ -151,9 +151,9 @@ public abstract class XdiAbstractClass<U extends XdiInstanceUnordered, O extends
 	 * @return The XDI element.
 	 */
 	@Override
-	public O getXdiInstanceOrdered(int index) {
+	public O getXdiInstanceOrdered(long index) {
 
-		XDI3SubSegment arcXri = XdiAbstractInstanceOrdered.createArcXri(Integer.toString(index));
+		XDI3SubSegment arcXri = XdiAbstractInstanceOrdered.createArcXri(Long.toString(index));
 
 		ContextNode contextNode = this.getContextNode().getContextNode(arcXri);
 		if (contextNode == null) return null;
@@ -175,7 +175,7 @@ public abstract class XdiAbstractClass<U extends XdiInstanceUnordered, O extends
 	 * Returns the number of XDI elements in this XDI class.
 	 */
 	@Override
-	public int getXdiInstancesOrderedCount() {
+	public long getXdiInstancesOrderedCount() {
 
 		return new IteratorCounter(this.getXdiInstancesOrdered()).count();
 	}
